feat: move onboarding to own router so we can set up apolloprovider
