+++ frontend/src/routes/index.tsx
@@ -1,21 +1,10 @@
 import { useRoutes } from 'react-router-dom'
 
 import { Chrome } from '../components/chrome'
-import {
-  Onboard,
-  OnboardHome,
-  OnboardNetwork,
-  OnboardSettings,
-  OnboardWalletCreate,
-  OnboardWalletImport
-} from '../components/onboard'
-import { Splash } from '../components/splash'
-import { useIsOnboard } from '../hooks/use-is-onboard'
 import { Home } from './home'
 import { Wallet } from './wallet'
 import { Sign } from './wallet/sign'
 import { WalletKeyPair } from './wallet/wallet-key-pair'
-import { WalletList } from './wallet/wallet-list'
 import { WalletCreate } from './wallet-create'
 import { WalletImport } from './wallet-import'
 
@@ -26,44 +15,8 @@ export enum Paths {
   Wallet = '/wallet'
 }
 
-// Nested paths DONT have '/'
-export enum OnboardPaths {
-  Settings = '/onboard/settings',
-  WalletCreate = '/onboard/wallet-create',
-  WalletImport = '/onboard/wallet-import',
-  Network = '/onboard/network'
-}
-
 export const AppRouter = () => {
-  const isOnboard = useIsOnboard()
-
   const routes = useRoutes([
-    {
-      path: Paths.Onboard,
-      element: <Onboard />,
-      children: [
-        {
-          index: true,
-          element: <OnboardHome />
-        },
-        {
-          path: OnboardPaths.Settings,
-          element: <OnboardSettings />
-        },
-        {
-          path: OnboardPaths.WalletCreate,
-          element: <OnboardWalletCreate />
-        },
-        {
-          path: OnboardPaths.WalletImport,
-          element: <OnboardWalletImport />
-        },
-        {
-          path: OnboardPaths.Network,
-          element: <OnboardNetwork />
-        }
-      ]
-    },
     {
       path: Paths.Wallet,
       element: <Wallet />,
@@ -107,11 +60,6 @@ export const AppRouter = () => {
       index: true
     }
   ])
-
-  // Wrap onboard pages with splash page
-  if (isOnboard) {
-    return <Splash>{routes}</Splash>
-  }
 
   // Rest of the pages get regular chrome with network
   return <Chrome>{routes}</Chrome>
